+++ popkorn-compiler/src/test/kotlin/cc/popkorn/compiler/PopKornCompilerTest.kt
@@ -1,7 +1,6 @@
 package cc.popkorn.compiler
 
-import cc.popkorn.PROVIDER_SUFFIX
-import cc.popkorn.RESOLVER_SUFFIX
+import cc.popkorn.compiler.utils.JavaClass
 import com.google.testing.compile.Compilation
 import com.google.testing.compile.CompilationSubject
 import com.google.testing.compile.Compiler
@@ -9,13 +8,19 @@ import com.google.testing.compile.JavaFileObjects
 import org.apache.commons.io.FileUtils
 import java.io.File
 
+
+/**
+ * Parent class of all compilation tests
+ *
+ * @author Pau Corbella
+ * @since 1.0.0
+ */
 abstract class PopKornCompilerTest {
-    private val packDash = "cc/popkorn/compiler"
     private val projectFolder = System.getProperty("user.dir")
     private val generatedFolder = "$projectFolder/build/generated/source/kaptKotlin/test"
 
 
-    fun assertCompileSuccess(vararg classes:ClassCreator){
+    fun assertCompileSuccess(vararg classes: JavaClass){
         try {
             val compilation = compile(*classes)
             CompilationSubject.assertThat(compilation).succeeded()
@@ -25,7 +30,7 @@ abstract class PopKornCompilerTest {
         }
     }
 
-    fun assertCompileFail(vararg classes:ClassCreator){
+    fun assertCompileFail(vararg classes: JavaClass){
         try{
             val compilation = compile(*classes)
             CompilationSubject.assertThat(compilation).failed()
@@ -38,8 +43,8 @@ abstract class PopKornCompilerTest {
 
 
 
-    private fun compile(vararg classes:ClassCreator) : Compilation{
-        val files = classes.map { JavaFileObjects.forSourceString(it.name(), it.construct()) }
+    private fun compile(vararg classes: JavaClass) : Compilation{
+        val files = classes.map { JavaFileObjects.forSourceString(it.qualifiedName(), it.construct()) }
 
         return Compiler.javac()
             .withProcessors(PopKornCompiler())
@@ -52,18 +57,11 @@ abstract class PopKornCompilerTest {
     }
 
 
-    private fun assertFiles(vararg classes:ClassCreator){
-        val listFiles = File("$generatedFolder/$packDash").listFiles()?.map { it.nameWithoutExtension } ?: throw RuntimeException("No generated files")
-        classes.map {
-            if (it.isPublicClass()){
-                "${it.simpleName()}_$PROVIDER_SUFFIX"
-            }else{
-                "${it.simpleName()}_$RESOLVER_SUFFIX"
-            }
-        }.forEach {
-            if (!listFiles.contains(it)) throw RuntimeException("Some providers/resolvers have not been generated")
-        }
-
+    private fun assertFiles(vararg classes: JavaClass){
+        classes.mapNotNull { it.getGeneratedFile()?.replace(".", "/") }
+            .map { File("$generatedFolder/$it.kt") }
+            .takeIf { it.all { it.exists() } }
+            ?: throw RuntimeException("Some providers/resolvers have not been generated")
     }