+++ popkorn-example/src/main/kotlin/cc/popkorn/example/model/D.kt
@@ -1,9 +1,7 @@
 package cc.popkorn.example.model
 
 import cc.popkorn.annotations.*
-import cc.popkorn.core.Provider
 import cc.popkorn.Scope
-import cc.popkorn.core.Injector
 
 @Injectable(alias = "is1") @ForEnvironments("env1") class D1: DiA, DiG
 
@@ -26,23 +24,23 @@ class D8: DiC
 class D10: Wrapper.DiD
 
 
-
-@InjectableProvider(alias = "is6")
+@InjectableProvider(scope = Scope.BY_NEW,alias = "is6")
 @ForEnvironments("env1", "env2", "env3")
-class DCustom6 : Provider<D6> {
-    override fun create(injector: Injector, environment: String?) = D6()
-    override fun scope() =  Scope.BY_NEW
+class DCustom6 {
+    fun create(d0:DiA) = D6()
+
+    @ForEnvironments("env2")
+    fun createEnv2() = D6()
+
 }
 
-@InjectableProvider(alias = "is7")
+@InjectableProvider(scope = Scope.BY_USE, alias = "is7")
 @ForEnvironments("env4")
-class DCustom7 : Provider<D7> {
-    override fun create(injector: Injector, environment: String?) = D7()
-    override fun scope() =  Scope.BY_USE
+class DCustom7 {
+    fun create() = D7()
 }
 
-@InjectableProvider(alias = "is8")
-class DCustom8 : Provider<D8> {
-    override fun create(injector: Injector, environment: String?) = D8()
-    override fun scope() =  Scope.BY_APP
+@InjectableProvider(scope = Scope.BY_APP, alias = "is8")
+class DCustom8 {
+    fun create() = D8()
 }