+++ popkorn/src/main/kotlin/cc/popkorn/annotations/Injectable.kt
@@ -5,7 +5,7 @@ import kotlin.reflect.KClass
 
 /**
  * Annotation to define that a class can be injected.
- * By default this annotation adds also as an injectable object all its interfaces.
+ * Also, if not defined, takes the scope BY_APP by default
  * If you want to exclude some of them, do it through parameter 'exclude'
  *
  * @author Pau Corbella