+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 Change Log
 ==========
 
-Version 1.2.0 *(UNKNOWN)*
+Version 1.2.0 *(2019-12-XX)*
 -----------------------------
 * New: Free `Injector` from it's internal modifier. Now can be used.
+* New: Now can use parametrized methods when injecting through `InjectableProvider`  
 
 Version 1.1.0 *(2019-11-29)*
 -----------------------------
-* New: Add PopKornCompat to use PopKorn from java code
+* New: Add `PopKornCompat` to use PopKorn from java code
 * New: Supports `internal` modifier for Injectable classes
 * New: Supports incremental annotation processing
 * New: Add support for Obfuscation. Only needs one rule: `-keep class * implements cc.popkorn.mapping.Mapping`